pyproject.toml
[project]
name = "deep-research"
version = "0.1.0"
description = "DeerFlow Agent Backend — LangGraph Lead Agent + FastAPI Gateway"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sse-starlette>=2.1.0",
    # LangChain / LangGraph
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.3.0",
    "langgraph-cli[inmem]>=0.1.71",
    "langchain-mcp-adapters>=0.1.0",
    # Configuration
    "pydantic>=2.12.5",
    "pyyaml>=6.0",
    "python-dotenv>=1.2.2",
    # HTTP
    "httpx>=0.27.0",
    # Tools
    "tavily-python>=0.7.0",
    # Legacy / kept for backward compat with medical_agent
    "elasticsearch>=9.3.0",
    "langchain-huggingface>=1.2.0",
    "sentence-transformers>=5.2.3",
    "sqlite-vec>=0.1.6",
]

[project.scripts]
gateway = "src.gateway.app:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.27.0",
]
